Refactor SearchActivity.java and ServiceDropdownAdapter.java
